Add rejection section
