+++ package.json
@@ -39,6 +39,7 @@
     "gatsby-plugin-react-helmet": "3.1.3",
     "gatsby-source-drupal-jsonapi-eci": "^0.1.0",
     "lodash": "4.17.15",
+    "mime": "2.4.4",
     "react": "16.9.0",
     "react-dom": "16.9.0",
     "react-helmet": "5.2.1",
