+++ static/_redirects
@@ -1,24 +1,24 @@
-/bg/initiatives/* /bg/initiatives/index.html 200
-/es/initiatives/* /es/initiatives/index.html 200
-/cs/initiatives/* /cs/initiatives/index.html 200
-/da/initiatives/* /da/initiatives/index.html 200
-/de/initiatives/* /de/initiatives/index.html 200
-/et/initiatives/* /et/initiatives/index.html 200
-/el/initiatives/* /el/initiatives/index.html 200
-/en/initiatives/* /en/initiatives/index.html 200
-/fr/initiatives/* /fr/initiatives/index.html 200
-/ga/initiatives/* /ga/initiatives/index.html 200
-/hr/initiatives/* /hr/initiatives/index.html 200
-/it/initiatives/* /it/initiatives/index.html 200
-/lv/initiatives/* /lv/initiatives/index.html 200
-/lt/initiatives/* /lt/initiatives/index.html 200
-/hu/initiatives/* /hu/initiatives/index.html 200
-/mt/initiatives/* /mt/initiatives/index.html 200
-/nl/initiatives/* /nl/initiatives/index.html 200
-/pl/initiatives/* /pl/initiatives/index.html 200
-/pt/initiatives/* /pt/initiatives/index.html 200
-/ro/initiatives/* /ro/initiatives/index.html 200
-/sk/initiatives/* /sk/initiatives/index.html 200
-/sl/initiatives/* /sl/initiatives/index.html 200
-/fi/initiatives/* /fi/initiatives/index.html 200
-/sv/initiatives/* /sv/initiatives/index.html 200
+/bg/* /bg/404.html 404
+/es/* /es/404.html 404
+/cs/* /cs/404.html 404
+/da/* /da/404.html 404
+/de/* /de/404.html 404
+/et/* /et/404.html 404
+/el/* /el/404.html 404
+/en/* /en/404.html 404
+/fr/* /fr/404.html 404
+/ga/* /ga/404.html 404
+/hr/* /hr/404.html 404
+/it/* /it/404.html 404
+/lv/* /lv/404.html 404
+/lt/* /lt/404.html 404
+/hu/* /hu/404.html 404
+/mt/* /mt/404.html 404
+/nl/* /nl/404.html 404
+/pl/* /pl/404.html 404
+/pt/* /pt/404.html 404
+/ro/* /ro/404.html 404
+/sk/* /sk/404.html 404
+/sl/* /sl/404.html 404
+/fi/* /fi/404.html 404
+/sv/* /sv/404.html 404
